list page with new chnages : table rearrangement
