+++ packages/manager/src/views/WebPropertyDetailPage/components/Applications/index.tsx
@@ -1,0 +1,1 @@
+export { Applications } from './ApplicationsListPage';
